default/requirements.txt
djangae==0.9.3
django==1.8
django-secure
django-csp
vk==2.0.2
requests==2.3
pylast==1.5.1
nose
six=1.10.0
google_api_python_client=1.5.0
httplib2=0.9.2
uritemplate=0.6
oauth2client=2.0.1
rsa=3.4
pyasn1_modules=0.0.8
pyasn1=0.1.9

# Uncomment these to use mapreduce or djangae.contrib.uniquetool (which uses mapreduce)
# These are forks of Google's libraries, which contain various fixes and patches
#git+https://github.com/potatolondon/appengine-pipelines.git
#git+https://github.com/potatolondon/potato-mapreduce.git#egg=mapreduce
